Consolidated BattleHonours endpoints under a single controller. Updated API endpoint mappings to follow standard conventions.
