+++ src/main/java/com/crusadehelper/controllers/CrusadeForceController.java
@@ -18,7 +18,7 @@ public class CrusadeForceController {
         this.service = service;
     }
 
-    @GetMapping("/crusadeforce/api/get")
+    @GetMapping("/api/CrusadeForce/v1/CrusadeForce")
     public ResponseEntity<String> getCrusadeForce(@RequestParam int crusadeId) throws JsonProcessingException {
         var cf = service.getCrusadeForceById(crusadeId);
         if(cf.isPresent())
@@ -26,7 +26,7 @@ public class CrusadeForceController {
         return ResponseEntity.notFound().build();
     }
 
-    @PostMapping("/crusadeforce/api/create")
+    @PostMapping("/api/CrusadeForce/v1/CrusadeForce")
     public ResponseEntity<String> createCrusadeForce() {
         var cf = service.createCrusadeForce();
         return ResponseEntity.ok("Create new crusade force with id: " + cf.getId());
